Working Buffer Serialization (FINALLY!)
Finally got vertex and index buffers to be correctly serialized and deserialized. 
Was originally trying to read buffer from BufferData class directly to buffer without resizing, this resulted in stack-based buffer overrun. This is because the size of the read class instance was larger than the place I was trying to read into. I solved this by instead creating a vector of type BufferData, then resizing the vector to the correct size before reading the data from binary.